first client open files to display tracks

+++ shared/src/main/scala/hu/mapro/mapping/Shared.scala
@@ -3,7 +3,7 @@ package hu.mapro.mapping
 import scala.concurrent.Future
 
 case class Position(lat: Double, lon: Double, timestamp: Long = 0)
-case class Track(positions: Seq[Position], id : Long = 0)
+case class Track(positions: Seq[Position], id : Int = 0)
 
 
 trait Api {
